Improve vents handling of filters
- The Vent log now supports unbounded length filter column for long
filters
- Changed interface for tags to simplify away ReflectTagged
- Optimised short-circuit path for tag filters (less reflection for
early results)
- Added errors to TxStack and BlockAccumulator where filters may give
invalid event stream WRT to tx formation
- Supported nested tag filters separated by '.' - e.g. 'Input.Address'
- Filter non-Log events from vent early
- Pull EventSpec and expose its fields to the query so now you can query
event name with `EventName = 'SolidityName'` since we have the ABIs. Other
fields in EventSpec are also exposed with the 'Event' prefix
- Add unit test for BlockConsumer function
