+++ event/query/tags.go
@@ -1,28 +1,17 @@
 package query
 
 import (
-	"sort"
+	"reflect"
+	"strings"
 )
 
+const defaultMaxTagDepth = 10
+
 type Tagged interface {
-	Keys() []string
 	Get(key string) (value interface{}, ok bool)
-	// Len returns the number of tags.
-	Len() int
 }
 
 type TagMap map[string]interface{}
-
-func MapFromTagged(tagged Tagged) map[string]interface{} {
-	tags := make(map[string]interface{})
-	for _, k := range tagged.Keys() {
-		v, ok := tagged.Get(k)
-		if ok {
-			tags[k] = v
-		}
-	}
-	return tags
-}
 
 func (ts TagMap) Get(key string) (value interface{}, ok bool) {
 	var vint interface{}
@@ -33,88 +22,72 @@ func (ts TagMap) Get(key string) (value interface{}, ok bool) {
 	return vint, true
 }
 
-func (ts TagMap) Len() int {
-	return len(ts)
+type CombinedTags []interface{}
+
+func TagsFor(vs ...interface{}) CombinedTags {
+	return vs
 }
 
-func (ts TagMap) Map() map[string]interface{} {
-	return ts
-}
-
-func (ts TagMap) Keys() []string {
-	keys := make([]string, 0, len(ts))
-	for k := range ts {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-	return keys
-}
-
-type CombinedTags struct {
-	keys []string
-	ks   map[string][]Tagged
-}
-
-func NewCombinedTags() *CombinedTags {
-	return &CombinedTags{
-		ks: make(map[string][]Tagged),
-	}
-}
-
-func MergeTags(tags ...Tagged) *CombinedTags {
-	ct := NewCombinedTags()
-	ct.AddTags(false, tags...)
-	return ct
-}
-
-func ConcatTags(tags ...Tagged) *CombinedTags {
-	ct := NewCombinedTags()
-	ct.AddTags(true, tags...)
-	return ct
-}
-
-// Adds each of tagsList to CombinedTags - choosing whether values for the same key should
-// be concatenated or whether the first should value should stick
-func (ct *CombinedTags) AddTags(concat bool, tagsList ...Tagged) {
-	for _, t := range tagsList {
-		for _, k := range t.Keys() {
-			if len(ct.ks[k]) == 0 {
-				ct.keys = append(ct.keys, k)
-				// Store reference to key-holder amongst Taggeds
-				ct.ks[k] = append(ct.ks[k], t)
-			} else if concat {
-				// Store additional tag reference only if concat, otherwise first key-holder wins
-				ct.ks[k] = append(ct.ks[k], t)
+func (ct CombinedTags) Get(key string) (interface{}, bool) {
+	for _, t := range ct {
+		tagged, ok := t.(Tagged)
+		if ok {
+			v, ok := tagged.Get(key)
+			if ok {
+				return v, true
 			}
 		}
-	}
-}
-
-func (ct *CombinedTags) Get(key string) (interface{}, bool) {
-	ts := ct.ks[key]
-	if len(ts) == 0 {
-		return "", false
-	}
-	values := make([]interface{}, 0, len(ts))
-	for _, t := range ts {
-		value, ok := t.Get(key)
+		v, ok := GetReflect(reflect.ValueOf(t), key)
 		if ok {
-			values = append(values, value)
+			return v, true
 		}
 	}
-	if len(values) == 0 {
+	return nil, false
+}
+
+func GetReflect(rv reflect.Value, key string) (interface{}, bool) {
+	return GetReflectDepth(rv, key, defaultMaxTagDepth)
+}
+
+var zeroValue = reflect.Value{}
+
+// Pull out values in a nested struct by following path
+func GetReflectDepth(rv reflect.Value, key string, maxDepth int) (interface{}, bool) {
+	if maxDepth < 0 {
 		return nil, false
 	}
-	if len(values) == 1 {
-		return values[0], true
+	if rv.Kind() == reflect.Ptr {
+		if rv.IsNil() {
+			return nil, false
+		}
+		rv = rv.Elem()
 	}
-	return values, true
+	keys := strings.SplitN(key, ".", 2)
+	field := rv.FieldByName(keys[0])
+	if field == zeroValue {
+		return nil, false
+	}
+	// If there there are unconsumed segments in the keys then descend
+	if len(keys) > 1 && (field.Kind() == reflect.Struct ||
+		field.Kind() == reflect.Ptr && field.Elem().Kind() == reflect.Struct) {
+		return GetReflectDepth(field, keys[1], maxDepth-1)
+
+	}
+	return field.Interface(), true
 }
 
-func (ct *CombinedTags) Len() (length int) {
-	return len(ct.keys)
+type taggedPrefix struct {
+	tagged Tagged
+	prefix string
 }
 
-func (ct *CombinedTags) Keys() []string {
-	return ct.keys
+func TaggedPrefix(prefix string, tagged Tagged) *taggedPrefix {
+	return &taggedPrefix{
+		prefix: prefix,
+		tagged: tagged,
+	}
+}
+
+func (t *taggedPrefix) Get(key string) (value interface{}, ok bool) {
+	return t.tagged.Get(strings.TrimPrefix(key, t.prefix))
 }
