+++ execution/exec/block_execution.go
@@ -2,6 +2,7 @@ package exec
 
 import (
 	"fmt"
+	"reflect"
 
 	"github.com/hyperledger/burrow/event"
 	"github.com/hyperledger/burrow/event/query"
@@ -48,23 +49,19 @@ func (be *BlockExecution) AppendTxs(tail ...*TxExecution) {
 }
 
 // Tags
-type TaggedBlockExecution struct {
-	query.Tagged
-	*BlockExecution
-}
 
-func (be *BlockExecution) Tagged() *TaggedBlockExecution {
-	return &TaggedBlockExecution{
-		Tagged: query.MergeTags(
-			query.TagMap{
-				event.EventIDKey:   EventStringBlockExecution(be.Height),
-				event.EventTypeKey: be.EventType(),
-			},
-			query.MustReflectTags(be),
-			query.MustReflectTags(be.Header),
-		),
-		BlockExecution: be,
+func (be *BlockExecution) Get(key string) (interface{}, bool) {
+	switch key {
+	case event.EventIDKey:
+		return EventStringBlockExecution(be.Height), true
+	case event.EventTypeKey:
+		return be.EventType(), true
 	}
+	v, ok := query.GetReflect(reflect.ValueOf(be.Header), key)
+	if ok {
+		return v, true
+	}
+	return query.GetReflect(reflect.ValueOf(be), key)
 }
 
 func QueryForBlockExecutionFromHeight(height uint64) *query.Builder {
@@ -73,46 +70,4 @@ func QueryForBlockExecutionFromHeight(height uint64) *query.Builder {
 
 func QueryForBlockExecution() *query.Builder {
 	return query.NewBuilder().AndEquals(event.EventTypeKey, TypeBlockExecution)
-}
-
-type TaggedBlockEvent struct {
-	query.Tagged
-	*StreamEvent
-}
-
-func (ev *StreamEvent) EventType() EventType {
-	switch {
-	case ev.BeginBlock != nil:
-		return TypeBeginBlock
-	case ev.BeginTx != nil:
-		return TypeBeginTx
-	case ev.Envelope != nil:
-		return TypeEnvelope
-	case ev.Event != nil:
-		return ev.Event.EventType()
-	case ev.EndTx != nil:
-		return TypeEndTx
-	case ev.EndBlock != nil:
-		return TypeEndBlock
-	}
-	return TypeUnknown
-}
-
-func (ev *StreamEvent) Tagged() *TaggedBlockEvent {
-	return &TaggedBlockEvent{
-		Tagged: query.MergeTags(
-			query.TagMap{
-				event.EventTypeKey: ev.EventType(),
-			},
-			query.MustReflectTags(ev.BeginBlock, "Height"),
-			query.MustReflectTags(ev.BeginBlock.GetHeader()),
-			query.MustReflectTags(ev.BeginTx),
-			query.MustReflectTags(ev.BeginTx.GetTxHeader()),
-			ev.Envelope.Tagged(),
-			ev.Event.Tagged(),
-			query.MustReflectTags(ev.EndTx),
-			query.MustReflectTags(ev.EndBlock, "Height"),
-		),
-		StreamEvent: ev,
-	}
 }
