+++ integration/rpctransact/call_test.go
@@ -359,7 +359,7 @@ func testCallTx(t *testing.T, kern *core.Kernel, cli rpctransact.TransactClient)
 			evAbi := spec.EventsByName["ChangeLevel"]
 			err = abi.UnpackEvent(&evAbi, log.Topics, log.Data, &direction, &depth)
 			require.NoError(t, err)
-			assert.Equal(t, evAbi.EventID.Bytes(), log.Topics[0].Bytes())
+			assert.Equal(t, evAbi.ID.Bytes(), log.Topics[0].Bytes())
 			assert.Equal(t, int64(18), depth)
 			assert.Equal(t, "Upsie!", direction)
 			return
@@ -381,7 +381,7 @@ func testCallTx(t *testing.T, kern *core.Kernel, cli rpctransact.TransactClient)
 			evAbi := spec.EventsByName["ManyTypes"]
 			data := abi.GetPackingTypes(evAbi.Inputs)
 			// Check signature
-			assert.Equal(t, evAbi.EventID.Bytes(), log.Topics[0].Bytes())
+			assert.Equal(t, evAbi.ID.Bytes(), log.Topics[0].Bytes())
 			err = abi.UnpackEvent(&evAbi, log.Topics, log.Data.Bytes(), data...)
 			require.NoError(t, err)
 
@@ -390,7 +390,7 @@ func testCallTx(t *testing.T, kern *core.Kernel, cli rpctransact.TransactClient)
 			expectedHash := h.Sum(nil)
 			// "Downsie!", true, "Donaudampfschifffahrtselektrizitätenhauptbetriebswerkbauunterbeamtengesellschaft", 102, 42, 'hash')
 			b := *data[0].(*[]byte)
-			assert.Equal(t, evAbi.EventID.Bytes(), log.Topics[0].Bytes())
+			assert.Equal(t, evAbi.ID.Bytes(), log.Topics[0].Bytes())
 			assert.Equal(t, "Downsie!", string(bytes.Trim(b, "\x00")))
 			assert.Equal(t, true, *data[1].(*bool))
 			assert.Equal(t, "Donaudampfschifffahrtselektrizitätenhauptbetriebswerkbauunterbeamtengesellschaft", *data[2].(*string))
