+++ vent/service/consumer_integration_test.go
